fix. service interface 삭제
