+++ src/main/java/com/example/gotogetherbe/chat/controller/ChatMemberController.java
@@ -1,6 +1,6 @@
 package com.example.gotogetherbe.chat.controller;
 
-import com.example.gotogetherbe.chat.service.impl.ChatMemberServiceImpl;
+import com.example.gotogetherbe.chat.service.ChatMemberService;
 import com.example.gotogetherbe.global.util.jwt.JwtUserDetails;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class ChatMemberController {
-  private final ChatMemberServiceImpl chatMemberService;
+  private final ChatMemberService chatMemberService;
 
   // 채팅방 참여자 목록 조회
   @GetMapping("/api/chat-member/list/{chatRoomId}")
