--- src/main/java/com/example/gotogetherbe/chat/service/impl/ChatRoomServiceImpl.java
@@ -1,114 +1,0 @@
-package com.example.gotogetherbe.chat.service.impl;
-
-import com.example.gotogetherbe.chat.entity.ChatMember;
-import com.example.gotogetherbe.chat.repository.ChatMemberRepository;
-import com.example.gotogetherbe.chat.dto.ChatRoomDto;
-import com.example.gotogetherbe.chat.entity.ChatRoom;
-import com.example.gotogetherbe.chat.repository.ChatRoomRepository;
-import com.example.gotogetherbe.chat.service.ChatRoomService;
-import com.example.gotogetherbe.chat.type.ChatRoomStatus;
-import com.example.gotogetherbe.global.exception.GlobalException;
-import com.example.gotogetherbe.global.exception.type.ErrorCode;
-import com.example.gotogetherbe.member.entitiy.Member;
-import com.example.gotogetherbe.member.repository.MemberRepository;
-import com.example.gotogetherbe.post.entity.Post;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class ChatRoomServiceImpl implements ChatRoomService {
-  private final ChatRoomRepository chatRoomRepository;
-  private final ChatMemberRepository chatMemberRepository;
-  private final MemberRepository memberRepository;
-  private final PostRepository postRepository;
-
-  @Override
-  @Transactional
-  public ChatRoomDto createChatRoom(String email, Long postId, String name) {
-    Member member = memberRepository.findByEmail(email)
-        .orElseThrow(() -> new GlobalException(ErrorCode.USER_NOT_FOUND));
-
-    Post post = postRepository.findById(postId)
-        .orElseThrow(() -> new GlobalException(ErrorCode.POST_NOT_FOUND));
-
-    if (!Objects.equals(member.getId(), post.getMember().getId())) {
-      throw new GlobalException(ErrorCode.MEMBER_POST_INCORRECT);
-    }
-
-    if (chatRoomRepository.existsById(postId)) {
-      throw new GlobalException(ErrorCode.ALREADY_CREATED_CHATROOM);
-    }
-
-    ChatRoom createdChatRoom = chatRoomRepository.save(ChatRoom.builder()
-            .post(post)
-            .name(name)
-            .status(ChatRoomStatus.ACTIVE)
-            .build());
-
-    return ChatRoomDto.from(createdChatRoom);
-  }
-
-  @Override
-  public List<ChatRoomDto> getChatRoomList(String email) {
-    Member member = memberRepository.findByEmail(email)
-        .orElseThrow(() -> new GlobalException(ErrorCode.USER_NOT_FOUND));
-
-    List<ChatMember> chatMemberList = chatMemberRepository.findAllByMemberId(member.getId());
-
-    List<ChatRoom> chatRoomList = chatMemberList.stream().map(ChatMember::getChatRoom).toList();
-
-    return chatRoomList.stream().map(ChatRoomDto::from).collect(Collectors.toList());
-  }
-
-  @Override
-  @Transactional
-  public String enterChatRoom(String email, Long chatRoomId) {
-    Member member = memberRepository.findByEmail(email)
-        .orElseThrow(() -> new GlobalException(ErrorCode.USER_NOT_FOUND));
-
-    ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
-        .orElseThrow(() -> new GlobalException(ErrorCode.CHATROOM_NOT_FOUND));
-
-    Optional<ChatMember> optionalChatMember = chatMemberRepository.findByChatRoomIdAndMemberId(chatRoom.getId(), member.getId());
-
-    if (optionalChatMember.isPresent()) {
-      throw new GlobalException(ErrorCode.ALREADY_PARTICIPANT_MEMBER);
-    }
-
-    ChatMember enterChatMember = chatMemberRepository.save(ChatMember.builder()
-        .chatRoom(chatRoom)
-        .member(member)
-        .build());
-
-    return enterChatMember.getMember().getNickname() + "님이 " + enterChatMember.getChatRoom().getName() + "에 참여하였습니다.";
-  }
-
-  @Override
-  @Transactional
-  public String exitChatRoom(String email, Long chatRoomId) {
-    Member member = memberRepository.findByEmail(email)
-        .orElseThrow(() -> new GlobalException(ErrorCode.USER_NOT_FOUND));
-
-    ChatRoom chatRoom = chatRoomRepository.findById(chatRoomId)
-        .orElseThrow(() -> new GlobalException(ErrorCode.CHATROOM_NOT_FOUND));
-
-    ChatMember chatMember = chatMemberRepository.findByChatRoomIdAndMemberId(chatRoom.getId(), member.getId())
-        .orElseThrow(() -> new GlobalException(ErrorCode.CHAT_MEMBER_NOT_FOUND));
-
-    chatMemberRepository.delete(chatMember);
-
-    List<ChatMember> chatMemberList = chatMemberRepository.findAllByChatRoomId(chatRoom.getId());
-
-    if (chatMemberList.isEmpty()) { // 채팅방 인원이 다 나갔다면 채팅방 삭제
-      chatRoomRepository.delete(chatRoom);
-    }
-
-    return chatMember.getMember().getNickname() + "님이 퇴장하였습니다.";
-  }
-}
